[Coroutines][5/6] Support 'opt -passes=coroutines'
Summary:
Depends on https://reviews.llvm.org/D71901.

The fifth in a series of patches that ports the LLVM coroutines passes
to the new pass manager infrastructure.

The first 4 patches allow users to run coroutine passes by invoking, for
example `opt -passes=coro-early`. However, most of LLVM's tests for
coroutines use an option, `opt -enable-coroutines`, which adds all 4
coroutines passes, in the correct order, to the legacy pass manager's pass
pipeline. This patch adds a similar feature, but in a way that can be used
by the new pass manager: a pipeline parser that understands
`opt -passes=coroutines`.

When used, this option schedules the passes 'coro-early', 'coro-split',
'coro-elide', 'coro-split', and 'coro-cleanup'.

Noteworthy here is that, in the old pass manager infrastructure, an indirect
call added by the 'coro-split' pass is then devirtualized by 'coro-elide',
and this would cause a restart of the CGSCC pass pipeline. The new pass
manager's CGSCC pass manager, on the other hand, does not check for
new and devirtualized function calls at the same intervals. Even using
the `DevirtSCCRepeatedPass` to wrap 'coro-split' and 'coro-elide', the check
for a devirtualized function would not detect a function added in the
first pass 'coro-split' and devirtualized in the second, 'coro-elide'.
So for now, this patch manually schedules iterations 1 and 2 of
'coro-split'.

Also worth mentioning: many of the coroutines regression tests check the
optimizations that occur with `opt -O2 -enable-coroutines`. With the new
pass manager, the opt driver does not honor arguments like `-O2`, which
add any number of passes. Instead, tests would need to list out passes
explicitly. I'm leaving that up to future passes, and for now, only
tests for `opt -O0 -enable-coroutines` have been modified to also test
the new pass manager.

Reviewers: GorNishanov, lewissbaker, chandlerc, junparser

Subscribers: EricWF, Prazek, hiraditya, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D71902