Introduce a CallGraph updater helper class
Summary:
The CallGraphUpdater is a helper that simplifies the process of updating
the call graph, both old and new style, while running an CG-SCC pass.

The uses are, for now, contained in different commits, e.g. D70767.

More functionality, e.g., replace a function with a new version, is
added as we need it.

Reviewers: chandlerc, modocache, hfinkel

Subscribers: wenlei, merge_guards_bot, mehdi_amini, mgorny, hiraditya, bollu, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D70927